gateway: emit integer created_at and completed_at on Responses envelopes
OpenAI emits both fields as integer epoch seconds. Strict typed clients
(Goose's Rust Responses decoder) reject the float the gateway used to
send, which made every gpt- alias unusable through Goose. The Python
and Rust Responses encoders now take integer seconds (floor); the
Chat Completions integer created is unchanged. Parity goldens and
fixtures move from fractional to whole seconds.

+++ exp/runtime/gateway/native/src/encode_responses.rs
@@ -141,7 +141,7 @@ pub struct ResponsesSseEncoder {
     response_id: String,
     synthetic_message_id: String,
     model: String,
-    created_at: f64,
+    created_at: i64,
     envelope: ResponsesEnvelope,
     started: bool,
     terminal: bool,
@@ -161,7 +161,7 @@ impl ResponsesSseEncoder {
     pub fn new(
         request_id: &str,
         model: &str,
-        created_at: f64,
+        created_at: i64,
         envelope: ResponsesEnvelope,
     ) -> Self {
         Self {
